+++ fx-application/fx-application-fxml/src/main/java/module-info.java
@@ -1,26 +1,27 @@
 // Copyright 2019 Axel Howind
-// 
+//
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 //     http://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-description = "JavaFX utilities (application)"
+module com.dua3.fx.application.fxml {
+    exports com.dua3.fx.application.fxml;
+    opens com.dua3.fx.application.fxml;
 
-javafx {
-    modules = listOf("javafx.base", "javafx.controls")
-}
+    requires transitive com.dua3.fx.application;
 
-dependencies {
-    api(project(":fx-util"))
-    api(project(":fx-controls"))
+    requires org.apache.logging.log4j;
 
-    implementation(rootProject.libs.dua3.utility)
+    requires static com.dua3.cabe.annotations;
+    requires com.dua3.utility;
+    requires javafx.fxml;
+    requires javafx.graphics;
 }