+++ fx-application/src/main/java/module-info.java
@@ -26,7 +26,6 @@ module com.dua3.fx.application {
     requires java.prefs;
     requires javafx.base;
     requires javafx.controls;
-    requires javafx.fxml;
     requires java.desktop;
     requires static com.dua3.cabe.annotations;
 }